chore: refactor move storage out to standalone dir
