+++ src/pages/User/content/MemberProfile.tsx
@@ -1,6 +1,6 @@
 import 'src/assets/css/slick.min.css'
 import type { IUserPP } from 'src/models/userPreciousPlastic.models'
-import type { IUploadedFileMeta } from 'src/stores/storage'
+import type { IUploadedFileMeta } from 'src/storage'
 import { Box, Image, Flex, Heading, Card, Paragraph } from 'theme-ui'
 import DefaultMemberImage from 'src/assets/images/default_member.svg'
 import { MemberBadge, UserStatistics, Username } from 'oa-components'
