+++ src/storage/index.tsx
@@ -5,6 +5,7 @@
 */
 
 import { storage } from '../utils/firebase'
+export type { IUploadedFileMeta } from './types'
 
 export class Storage {
   /****************************************************************************** *
@@ -54,18 +55,4 @@ const generateUploadedFileMeta = (
     updated: snapshotMeta.updated,
   }
   return fileInfo
-}
-
-/****************************************************************************** *
-        Interfaces
-/****************************************************************************** */
-export interface IUploadedFileMeta {
-  downloadUrl: string
-  contentType?: string | null
-  fullPath: string
-  name: string
-  type: string
-  size: number
-  timeCreated: string
-  updated: string
 }
