+++ src/storage/types.ts
@@ -1,0 +1,13 @@
+/****************************************************************************** *
+        Interfaces
+/****************************************************************************** */
+export interface IUploadedFileMeta {
+  downloadUrl: string
+  contentType?: string | null
+  fullPath: string
+  name: string
+  type: string
+  size: number
+  timeCreated: string
+  updated: string
+}
