+++ src/stores/Maps/maps.store.ts
@@ -14,7 +14,7 @@ import { getUserAvatar } from '../User/user.store'
 import { MAP_GROUPINGS } from './maps.groupings'
 import { generatePins, generatePinDetails } from 'src/stores/Maps/generatePins'
 import type { IUserPP } from 'src/models/userPreciousPlastic.models'
-import type { IUploadedFileMeta } from '../storage'
+import type { IUploadedFileMeta } from '../../storage'
 import {
   hasAdminRights,
   needsModeration,
