+++ src/stores/common/module.store.ts
@@ -5,13 +5,13 @@ import { logger } from 'src/logger'
 import { includesAll } from 'src/utils/filters'
 import { stripSpecialCharacters } from 'src/utils/helpers'
 
-import { Storage } from '../storage'
+import { Storage } from '../../storage'
 
 import type { ISelectedTags } from 'src/models/tags.model'
 import type { IDBEndpoint, ILocation } from 'src/models/common.models'
 import type { RootStore } from '../index'
 import type { IConvertedFileMeta } from '../../types'
-import type { IUploadedFileMeta } from '../storage'
+import type { IUploadedFileMeta } from '../../storage'
 /**
  * The module store is used to share methods and data between other stores, including
  * `db` - the common database
